Adds a course option to auto-advance videos.
If enabled for a course, as soon as the video ends, the next unit or subsection
will be loaded, and if it contains a single video, that video will be played.

Course authors can enable the setting for a course, but learners can toggle the
setting on or off once it's enabled on the course.

(cherry picked from commit ecf01d1b522c2dbb4a912e0f05c1a5f2ec2911d4)