language --> content_language and get_fields() actually returns
Upstream PR: https://github.com/edx/edx-platform/pull/16068

This was required in order to call `get_course_run_details`
inside of the course about template.

In particular, this PR is a follow-up to the previous
commit here, because the other PR had some blatant errors but
still managed to get merged. >_>