calculated hours of effort for rendering cert template context
Upstream PR: https://github.com/edx/edx-platform/pull/16088

This was required in order to call `get_course_run_details`
inside of the course about template.

In particular, this PR is the *2nd* follow-up to the previous
commit here, because the other PR had some blatant errors but
still managed to get merged. >_>