[BB-112] Fix the order randomization behaviour of Randomized Content Block
cf https://github.com/edx/edx-platform/pull/24343

Fix the order randomization behaviour of Randomized Content Block

The Randomized Content Block XBlock only randomizes the selection of
the children blocks and has unpredictable randomization of
the order of the selected child blocks due to the usage of sets, which
are unordered, for storing the selected blocks. This becomes apparent
when all the available child blocks in a library are chosen for a
Randomized Content Block, to randomize just the order of the child
blocks and not just the selection of the blocks. The order of the
selected blocks ends up being similar for multiple learners.

This change modifies the XBlock to store the selected child blocks in
a list, instead of a set, after randomly shuffling them. It also adds
a new block transformer 'ContentLibraryTransformer', which randomizes
the order of the selected blocks.

(cherry picked from commit 967688f86718ebd7fa9b53ac05a482e4b4dc5782)

Address review comment

(cherry picked from commit 2fab6dc8954f214a279d3ab6752f93470ffd2ebf)