Fixes bug with field values in Studio edit view
Fields like "rerandomize" and "showanswer" were not showing these fields' list
of available values when editing the XBlock settings.  The cause has to do with
how the XBlock's list of fields is generated.

XBlock.ScopedStorageMixin treats the base class as special -- it can override any of
the subclass's field specifications (and their options).

However, since the capa_base.CapaField class was just a mixin and not the
actual base class, the fields brought in by the InheritanceMixin were taking
precedence over the CapaField fields.

Solution: move the CapaXBlock fields out of CapaFields and into CapaXBlock.