Allows the CapaXBlock to determine its own Studio edit modal height
CapaXBlock, like ORA2, and all StudioEditable XBlocks, display their own tabs
and buttons.  The platform recognizes this, and in response, hides the header
and button bars.

However, because the modal content box remained a fixed height, this made any
dynamic height management by XBlocks for their own header and button bars
basically impossible.

(Note that ORA2 gets around this by adding the "openassessment_full_height"
class to div.modal-content, which sets its height to 100%.)

This change moves the "hide header and button bar" out of JS and into CSS, and
also changes the height of the modal content box to "auto", so that the XBlocks
can determine it themselves.