Updates jake Clean task and BuildTools Uglify options
Jake clean is updated in order to not delete the plugin.min.js for 'image' and 'media' plugins because these plugins do not have a 'plugin.js' to create the minified file from.

Also removes 'outputMinified' for the 'minify' tasks that are included with the 'minify-plugin' so that the minification is done through uglify instead (to keep things uniform and consistent).