Adds waffle switch to enable third party auth email association
When logging in with an SSO provider, there are multiple cases that result in the confusing warning message that requires you to login in again in roder to link your edX account and your SSO provider account together.

This functionality is often not wanted. Often what is needed is a one-to-one association between the edX account and the SSO account, unless linked differently from the account settings.

Therefore, this change adds a waffle switch in order to associate the edX account and SSO account by email on login, through the social auth pipeline.