fix: set answer notification alignment to bottom (#27022)
On focus to the answer submition notification, the
notification was scrolled to center of screen. This
behaviour was on Chrome(87+) browser.
This commits overrides the focus jQuery plugin to
set the alignment of answer notifications to the bottom of the
viewport.

Fix the custom focus jQuery plugin

Co-authored-by: Agrendalath
(cherry picked from commit 52ab785764b4f1edd33a574dd8979e4f23f97867)