fix: hide unavailable sequence title in course breadcrumbs (#28028)
When a sequence title was unavailable for a user (e.g. before starting a timed
exam), the `None` value was displayed in course breadcrumbs. This hides it in
such cases and ensures that a trailing breadcrumb separator is not displayed.

(cherry picked from commit a5d9dd4fb9f4264626776c67c4013a7687214669)