feat: add reset option to Randomized Content Block
This adds a reset option to a Randomized Content Block (LibraryContentBlock)
to allow an OpenEdx user to reset questions in a Randomized Content Block.
This is achieved by adding a `Reset Problems` button to the UI of a vertical
with randomized content, and a reset handler to process the request.