fix: update lti-consumer-xblock to 2.10.1 (#400)
The previous lti-consumer-xblock commit hash was pointing to a code drift branch which was backporting https://github.com/edx/xblock-lti-consumer/pull/150 to tag `2.4.0`

Unfortunately, `config_id` duplicate errors started showing for one of the clients. Accordingly, we had to update the xblock to the latest version which inculded that fix and still supported koa (`2.10.1`).

However, https://github.com/edx/xblock-lti-consumer/pull/150 was not merged until tag `3.0.3`. Thefore, we created a custom tag, `2.10.1-opencraft` which includes the backported change on top of tag `2.10.1`