fix: patch security issues in studio's xblock_handler api
It is possible for a non-course-authors to alter the content of xblocks
which use the StudioEditableXBlockMixin. The vulnerability at a minimum
would allow a non-privileged user to achieve HTML injection in some
course content.

No checks for permission course author privilege was made on the
xblock_handler api. Now to submit edits from studio in this way, a user
has to be an author.

Internal-ref: https://tasks.opencraft.com/browse/SE-5795