fix: apply language specified in Site Configuration
Previously, this solution was implemented in the `dark_lang` middleware.
However, the `lang_pref` middleware started setting the user-preferred
language by default, which resulted in overriding the language cookie twice
for a single request (making this feature unusable). In this commit, the
language from Site Configuration will be applied after the one specified in
user preferences, to allow setting a language globally for a specific Site.