feat: allow disabling spaced out sections in self paced courses
In self paced courses, if relative due dates are enabled via
SelfPacedRelativeDatesConfig, all graded content would be assigned
relative due dates which are evenly spaced out over an estimated
duration of a course (aka. Personal Learner Schedule or PLS). If
CUSTOM_RELATIVE_DATES are enabled, custom set relative due dates would
(sometimes) override the "spaced out" ones.

However, there are some usecases, when custom relative due dates are
desired, without the PLS. For this usecase we are adding a
DISABLE_SPACED_OUT_SECTIONS CourseWaffleFlag. None of the existing
behaviour is changed unwillingly. When the flag is enabled, the relative
due dates will only be applied to the subsections that have custom
relative due dates set, or when a similar setting is set in Advanced
Settings of a course.